narvar/analysis_peak_2025_sonnet45/scripts/requirements.txt
google-cloud-bigquery>=3.10.0

